# specs/apps/crud/contracts/generated/openapi-bundled.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crud API",
    "description": "REST API contract for the polyglot crud expense tracker crud application",
    "version": "1.0.0",
    "contact": {
      "name": "ose-primer",
      "url": "https://github.com/wahidyankf/ose-primer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default backend (Go/Gin)"
    },
    {
      "url": "http://localhost:8201",
      "description": "Java Spring Boot backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "List all users (admin)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized (not admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/disable": {
      "post": {
        "operationId": "disableUser",
        "summary": "Disable a user account",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User disabled"
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/enable": {
      "post": {
        "operationId": "enableUser",
        "summary": "Enable a disabled user account",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User enabled"
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/unlock": {
      "post": {
        "operationId": "unlockUser",
        "summary": "Unlock a locked user account",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unlocked"
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/force-password-reset": {
      "post": {
        "operationId": "forcePasswordReset",
        "summary": "Force password reset for a user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset forced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "description": "User account details",
        "required": [
          "id",
          "username",
          "email",
          "displayName",
          "status",
          "roles",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DISABLED",
              "LOCKED"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DisableRequest": {
        "type": "object",
        "description": "Request to disable a user account",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "PasswordResetResponse": {
        "type": "object",
        "description": "Response containing a temporary password reset token",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "UserListResponse": {
        "type": "object",
        "description": "Paginated list of users (admin view)",
        "required": [
          "content",
          "totalElements",
          "totalPages",
          "page",
          "size"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "totalElements": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standardized error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}